requirements.txt
python-telegram-bot[ext]==21.3
python-dotenv==1.0.1
httpx~=0.27.0
openai==1.37.1
numpy==1.26.4
pandas==2.2.2
